Implemented stub for mblen() that simply calls strlen(). Probably it's extremely broken, but since the manual states that the result of mblen() depends on LC_CTYPE (which we can say is always "ISO8859-1" on AROS) I hope it's not to bad...
git-svn-id: https://svn.aros.org/svn/aros/trunk@21597 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/include/stdlib.h
@@ -151,7 +151,7 @@ lldiv_t lldiv(long long int numer, long long int denom);
 #endif
 
 /* Multibyte character functions */
-int mblen(const char *s, size_t n);
+size_t mblen(const char *s, size_t n);
 int mbtowc(wchar_t * restrict pwc, const char * restrict s, size_t n);
 int wctomb(char *s, wchar_t wchar);
